# feat: Optimize API performance with 84% speed improvement