… OPTIMIZATION_PLAN.md
… PERFORMANCE_RESULTS.md
… prisma/schema.prisma
… src/app/api/scans/[id]/route.ts
… src/app/image/[name]/scan/[scanId]/page.tsx
… src/lib/scanner/DatabaseAdapter.ts